Move styles to seperate file
